@@ -140,11 +137,20 @@ val ktlintFormat by tasks.creating(JavaExec::class) {
 	jvmArgs("--add-opens", "java.base/java.lang=ALL-UNNAMED")
 }
 
-tasks.withType<KotlinCompile> {
-	kotlinOptions {
-		freeCompilerArgs = listOf("-Xjsr305=strict")
-		jvmTarget = "16"
-	}
+tasks.withType<KotlinCompile>() {
+
+    kotlinOptions {
+        jvmTarget = "21"
+        apiVersion = "2.1"
+        languageVersion = "2.1"
+        freeCompilerArgs = listOf("-Xjsr305=strict")
+    }
+
+    // you can also add additional compiler args,
+    // like opting in to experimental features
+    kotlinOptions.freeCompilerArgs += listOf(
+        "-opt-in=kotlin.RequiresOptIn",
+    )
 }
 
 jib {
