@@ -7,54 +7,52 @@ import io.swagger.v3.oas.annotations.enums.SecuritySchemeType
 import io.swagger.v3.oas.annotations.security.SecurityScheme
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.Import
+
 import org.springframework.http.HttpMethod
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
-import org.springframework.security.config.annotation.web.builders.WebSecurity
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter
+import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer
 import org.springframework.security.config.http.SessionCreationPolicy
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
 import org.springframework.security.crypto.password.PasswordEncoder
+import org.springframework.security.web.DefaultSecurityFilterChain
 import org.springframework.security.web.authentication.AnonymousAuthenticationFilter
 import org.springframework.security.web.header.writers.ReferrerPolicyHeaderWriter
-import org.springframework.web.filter.CorsFilter
+import org.springframework.security.web.util.matcher.AntPathRequestMatcher
+import org.springframework.security.web.util.matcher.AntPathRequestMatcher.antMatcher
 import org.springframework.web.filter.ForwardedHeaderFilter
-import org.zalando.problem.spring.web.advice.security.SecurityProblemSupport
+
 
 @Configuration
 @EnableWebSecurity(debug = true)
-@EnableGlobalMethodSecurity(prePostEnabled = true, securedEnabled = true)
-@Import(SecurityProblemSupport::class)
+@EnableMethodSecurity(securedEnabled = true, jsr250Enabled = true)
 @SecurityScheme(
     name = "jwt_auth", type = SecuritySchemeType.HTTP, bearerFormat = "JWT", scheme = "bearer"
 )
 class SecurityConfiguration(
     private val tokenProvider: TokenProvider,
-    private val corsFilter: CorsFilter,
-    private val problemSupport: SecurityProblemSupport,
-) : WebSecurityConfigurerAdapter() {
-
+) {
     @Bean
     fun forwardedHeaderFilter(): ForwardedHeaderFilter? {
         return ForwardedHeaderFilter()
     }
 
-    override fun configure(web: WebSecurity) {
-
-        web
-            .ignoring()
-            .antMatchers(HttpMethod.OPTIONS, "/**")
-            .antMatchers("/app/**/*.{js,html}")
-            .antMatchers("/i18n/**")
-            .antMatchers("/content/**")
-            .antMatchers("/swagger-ui/**")
-            .antMatchers("/test/**")
+    @Bean
+    fun webSecurityCustomizer(): WebSecurityCustomizer {
+        return WebSecurityCustomizer {
+            it
+                .ignoring()
+                .requestMatchers(antMatcher(HttpMethod.OPTIONS,"/**"))
+                .requestMatchers(antMatcher("/app/**/*.{js,html}"))
+                .requestMatchers(antMatcher("/i18n/**"))
+                .requestMatchers(antMatcher("/content/**"))
+                .requestMatchers(antMatcher("/swagger-ui/**"))
+        }
     }
 
-    @Throws(Exception::class)
-    override fun configure(http: HttpSecurity) {
+    @Bean
+    fun securityFilterChain(http: HttpSecurity): DefaultSecurityFilterChain {
         /* ktlint-disable max_line_length */
         // @formatter:off
         http
@@ -62,10 +60,6 @@ class SecurityConfiguration(
             .and()
             .csrf()
             .disable()
-            .exceptionHandling()
-            .authenticationEntryPoint(problemSupport)
-            .accessDeniedHandler(problemSupport)
-            .and()
             .headers()
             .contentSecurityPolicy("default-src 'self'; frame-src 'self' data:; script-src 'self' 'unsafe-inline' 'unsafe-eval' https://storage.googleapis.com; style-src 'self' 'unsafe-inline'; img-src 'self' data:; font-src 'self' data:") /* ktlint-ignore */
             .and()
@@ -77,22 +71,25 @@ class SecurityConfiguration(
             .sessionManagement()
             .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
             .and()
-            .authorizeRequests()
-            .antMatchers("/api/account/reset-password/init").permitAll()
-            .antMatchers("/api/account/reset-password/finish").permitAll()
-            .antMatchers("/api/admin/**").hasAuthority(AuthoritiesConstants.ADMIN)
-            .antMatchers("/api/**").permitAll()
-            .antMatchers("/management/health").permitAll()
-            .antMatchers("/management/health/**").permitAll()
-            .antMatchers("/management/info").permitAll()
-            .antMatchers("/management/prometheus").permitAll()
-            .antMatchers("/management/**").hasAuthority(AuthoritiesConstants.ADMIN)
-            .and().addFilterAfter(
+            .authorizeHttpRequests{ requests ->
+                requests
+                    .requestMatchers(AntPathRequestMatcher("/api/account/reset-password/init")).permitAll()
+                    .requestMatchers(AntPathRequestMatcher("/api/account/reset-password/finish")).permitAll()
+                    .requestMatchers(AntPathRequestMatcher("/api/admin/**")).hasAuthority(AuthoritiesConstants.ADMIN)
+                    .requestMatchers(AntPathRequestMatcher("/api/**")).hasAuthority(AuthoritiesConstants.USER)
+                    .requestMatchers(AntPathRequestMatcher("/management/health")).permitAll()
+                    .requestMatchers(AntPathRequestMatcher("/management/health/**")).permitAll()
+                    .requestMatchers(AntPathRequestMatcher("/management/info")).permitAll()
+                    .requestMatchers(AntPathRequestMatcher("/management/prometheus")).permitAll()
+                    .requestMatchers(AntPathRequestMatcher("/management/**")).hasAuthority(AuthoritiesConstants.ADMIN)
+            }
+            .addFilterAfter(
                 JwtFilter(tokenProvider), AnonymousAuthenticationFilter::class.java
             )
         http.headers().cacheControl()
         // @formatter:on
         /* ktlint-enable max_line_length */
+        return http.build()
     }
 
     @Bean
