@@ -12,6 +12,7 @@ import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.ComponentScan
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories
 import org.springframework.web.servlet.config.annotation.PathMatchConfigurer
+import org.springframework.web.util.pattern.PathPatternParser
 
 @SpringBootApplication(exclude = [ErrorMvcAutoConfiguration::class])
 
@@ -43,7 +44,4 @@ fun main(args: Array<String>) {
     runApplication<LostcitiesMatchesApplication>(*args)
 }
 
-@Bean
-fun configureMatcher(configurer: PathMatchConfigurer) {
-    configurer.setUseTrailingSlashMatch(false)
-}
+
