@@ -1,9 +1,12 @@
 package io.dereknelson.lostcities.matches.persistence
 
 import org.springframework.data.domain.Page
+import org.springframework.data.domain.PageRequest
 import org.springframework.data.domain.Pageable
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
+import org.springframework.data.repository.PagingAndSortingRepository
+import org.springframework.data.repository.query.Param
 import org.springframework.stereotype.Repository
 
 @Repository
@@ -22,32 +25,28 @@ interface MatchRepository : JpaRepository<MatchEntity, Long> {
         (
             matchEntity.player1 = :playerName OR matchEntity.player2 = :playerName
         ) AND
-        matchEntity.isReady IS TRUE
+        matchEntity.isReady = TRUE
     ORDER BY matchEntity.lastModifiedDate
         """
     )
     fun findActiveMatches(playerName: String, page: Pageable): Page<MatchEntity>
 
-    @Query(
-        """
-    SELECT matchEntity
+    @Query(value = """SELECT matchEntity
     FROM MatchEntity matchEntity
     WHERE
-        (matchEntity.player1 <> :playerName AND matchEntity.player2 IS NULL) AND
-        matchEntity.isReady IS FALSE AND
-        matchEntity.isCompleted IS FALSE
+        matchEntity.player1 <> :playerName AND
+        matchEntity.player2 = null AND
+        matchEntity.isReady = false AND
+        matchEntity.isCompleted = false
+    """)
+    fun findAvailableMatches(@Param("playerName") playerName: String, page: Pageable): Page<MatchEntity>
 
-        """
-    )
-    fun findAvailableMatches(playerName: String, page: Pageable): Page<MatchEntity>
-
-    @Query(
-        """
-    SELECT matchEntity
+    @Query( value = """
+        SELECT matchEntity
     FROM MatchEntity matchEntity
     WHERE
         ( matchEntity.player1 = :playerName OR matchEntity.player2 = :playerName ) AND
-        matchEntity.isCompleted IS TRUE
+        matchEntity.isCompleted = TRUE
         """
     )
     fun findCompletedMatches(playerName: String, page: Pageable): Page<MatchEntity>
